fixes for rendered documentation of utils.color.*
