+++ docs/source/reference_index/utilities_misc.rst
@@ -11,7 +11,12 @@ Module Index
 
    constants
    ~utils.bezier
-   ~utils.color
+   ~utils.color.core
+   ~utils.color.manim_colors
+   ~utils.color.AS2700
+   ~utils.color.BS381
+   ~utils.color.XKCD
+   ~utils.color.X11
    ~utils.commands
    ~utils.config_ops
    ~utils.deprecation
